fix(agents): settle CLI bootstrap after cleanup
