+++ src/agents/command/assistant-transcript-repair.test.ts
@@ -256,4 +256,3 @@ describe("repairPendingAssistantTranscriptTurns", () => {
     expect(sessionStore[sessionKey]?.sessionId).toBe("replacement");
   });
 });
-
