+++ src/agents/embedded-agent-runner/run/attempt.spawn-workspace.context-injection.test.ts
@@ -4,17 +4,14 @@ import { beforeEach, describe, expect, it, vi } from "vitest";
 import { filterHeartbeatTranscriptArtifacts } from "../../../auto-reply/heartbeat-filter.js";
 import { HEARTBEAT_PROMPT } from "../../../auto-reply/heartbeat.js";
 import type { BootstrapContextRunKind } from "../../bootstrap-mode.js";
-import { assembleHarnessContextEngine } from "../../harness/context-engine-lifecycle.js";
 import {
   isPrimaryInteractiveBootstrapRun,
   resolveBootstrapContextInjection,
 } from "../../bootstrap-routing.js";
+import { assembleHarnessContextEngine } from "../../harness/context-engine-lifecycle.js";
 import { limitHistoryTurns } from "../history.js";
 import { buildEmbeddedMessageActionDiscoveryInput } from "../message-action-discovery-input.js";
-import {
-  type AttemptContextEngine,
-  resolveAttemptBootstrapContext,
-} from "./attempt-context-engine-helpers.js";
+import type { AttemptContextEngine } from "./attempt-context-engine-helpers.js";
 import { resetEmbeddedAttemptHarness } from "./attempt-spawn-workspace.test-support.js";
 
 async function resolveBootstrapContext(params: {
